polymarket_hft_bot/requirements.txt
aiohttp>=3.10,<4
python-dotenv>=1.0,<2
websockets>=13,<15
py-clob-client>=0.25,<1
